requirements.txt
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
httpie==1.0.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
